requirements.txt
# Core AI Frameworks
google-adk>=0.3.0
crewai>=0.55.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
openai>=1.40.0

# Vector Store
chromadb>=0.5.0

# Search & Scraping
tavily-python>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Resume Parsing
PyPDF2>=3.0.0
python-docx>=1.1.0

# API & Server
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Storage
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# Observability
langsmith>=0.1.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.12.0
